continue debugging

+++ headers.h
@@ -74,7 +74,7 @@ struct arp_header {
     uint8_t protocol_size;
     uint16_t opcode;
     uint8_t sha[MAC_LEN];
-    uint8_t spa[IP_LEN];
+    uint32_t spa[IP_LEN];
     uint8_t tha[MAC_LEN];
     uint8_t tpa[IP_LEN];
 }__attribute__((packed));
